refactor: extract logic from index.js
